Update SetAppointment.jsx
Updated SetAppointment.jsx by fixing the spaces of checkbox sentence and fixing the grammar in header in step 5 (Confirmation Section).

Fixed modal to not be closed when esc key is pressed. Also fixed modal by making the webpage reload after the user clicks the download button.

This is all in step 5.