Update Purpose.jsx
Updated Purpose.jsx by removing --Select Purpose-- as default value in the Dropdowns. This was causing a bug. Example, adding purpose (dropdown) works. However, clicking the default value (--Select Purpose--) strangely deletes the dropdown itself. Hence, I removed the default value (--Select Purpose--) and made it as a placeholder for the dropdown instead. So that the dropdown won't consider it as a real value and won't cause that strange bug.