Update TimePicker.jsx
Updated TimePicker.jsx by adding props and function to fix the issue of disabled time being selected. Example, user selects nov 14 and its timeslot 8:00 until it exceeded the amount of appointments (8:00 will be disabled). However, since the 8:00 on nov 14 is finally disabled. Users can still access this time slot by clicking other dats like nov 15 then selecting 8:00. Doing that, then clicking the nov 14 again selects the disabled 8:00 timeslot of that nov 14.

To solve that, I created a function and logic to move the selected time to available timeslots if a user for example selects nov 15 8:00. Then clicks nov 14 again, but this time the selected time will go to 9:00 or other available time because 8:00 is already full.