Create ProtectedStudentRoute.jsx
Created ProtectedStudentRoute.jsx to provide a simple authentication to student account based.

Note: (ProtectedStudentRoute.jsx) and (optionally the useStudentAuth.jsx) can be stored all inside the other two files (ProtectedRoute.jsx) and (useAuth.jsx).

The newly created two are for users. While the last two are for admins. However, it can be placed in one file. But to be distinct, this works as well.

This will be decided and updated in the future.